% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\docType{data}
\name{argument_reference}
\alias{argument_reference}
\title{Attributes}
\format{
An object of class \code{Argument_Reference} of length 1.
}
\usage{
argument_reference
}
\arguments{
\item{argument_reference}{object containing information about input models 
and parameters for cft.}
}
\description{
Attributes
}
\keyword{datasets}
