% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_data.R
\name{available_data}
\alias{available_data}
\title{Available data}
\usage{
available_data(
  web_link = "https://cida.usgs.gov/thredds/dodsC/macav2metdata_daily_future",
  verbose = TRUE
)
}
\arguments{
\item{web_link}{}

\item{verbose}{}
}
\value{

}
\description{
Available data
}
\examples{

}
