% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cftdata.R
\name{cftdata}
\alias{cftdata}
\title{Climate Futures Toolbox Data}
\usage{
cftdata(
  shp_path,
  area_name,
  park,
  dataset = "maca",
  models = get_reference("maca")$models,
  parameters = get_reference("maca")$parameters,
  scenarios = get_reference("maca")$scenarios,
  years = c(1950, 2099),
  project_dir = tempdir(),
  verbose = TRUE,
  ncores = 1
)
}
\arguments{
\item{shp_path}{A path to a shapefile with which to clip the resulting data
sets. This path may be to a local .shp file or a url to a zipped remote
file. If this option is used, leave the `park` argument empty or set
to NA. (character)}

\item{area_name}{If a shapefile path is used, provide a name to use as a
reference to the location. This will be used in file names, attributes, and
directories. (character)'}

\item{park}{The name of a national park (e.g., "Yellowstone National
Park", "Yellowstone Park", "Yellowstone", "yellowstone", etc.). The user may
use this option in place of a shapefile path. In this case, leave the
shp_path argument empty or set to NA. (character)}

\item{dataset}{The name of the data source from which to retrieve data. Defaults
to "maca". (character)
 \itemize{
   \item "maca" = Multivariate Adaptive Constructed Analogs
   \item "gridmet" = Gridded Surface Meteorological Dataset
 }}

\item{models}{A list of global circulation models to download. Only available
for GCMs. If left empty all available models will be downloaded. A list of
currently available models is available under cft::get_reference("maca")$models.
(vector)}

\item{parameters}{A list of climate parameters to download. If left empty
all available parameters will be downloaded. A list of available of models is
available under cft::get_reference(<dataset>)$parameters. (vector)}

\item{scenarios}{A list of representative concentration pathways (rcps) to
download. If left empty all available rcps will be downloaded. A list of
available of rcps is available under cft::get_reference(<dataset>)$scenarios.
(vector)}

\item{years}{The first and last years of the desired period. (vector)}

\item{project_dir}{The local directory in which to save files. By default,
files are saved in a temporary directory (as per CRAN guidelines), and are
lost after your R session ends. Specify a path to a local directory with
this argument to retain files and avoid duplicate downloads
in subsequent R sessions.  (character)}

\item{verbose}{Print verbose output. (logical)}

\item{ncores}{The number of cpus to use, which defaults to 1. (numeric)}
}
\value{
A tibble containing information about climate data files.
}
\description{
Retrieves subsets of historical or projected climate datsets.
Available datasets include a Multivariate Adaptive Constructed Analogs (MACA)
dataset and the Gridded Surface Meteorological Dataset (gridMET).
}
\details{
This package retrieves daily gridded data sets of General Circulation Model  # We need to rephrase this
(GCM) runs clipped to areas of interest and returns a data frame of the
file names and they're storage paths. Each of these data sets represent
a single GCM, climate variable and Representative Concentration Pathway (RCP)
from 1950 to 2099. The 1950 to 2005 portion of this time period represents
historical data while the 2006 to 2099 portion represents modeled data.
The original data sets may be found at
\url{http://thredds.northwestknowledge.net:8080/thredds/reacch_climate_CMIP5_aggregated_macav2_catalog.html}
}
\examples{
\donttest{
d <- cftdata(park = "Acadia National Park", parameters = "pr", dataset = "maca",
             years = c(2020, 2021), models = "CCSM4", scenarios = "rcp85",
             ncores = parallel::detectCores())
}

}
