% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarityIndicesFunctions.R
\name{createDissimilarityIndicesEstimator}
\alias{createDissimilarityIndicesEstimator}
\title{Instantiate an Estimator of Dissimilarity Indices}
\usage{
createDissimilarityIndicesEstimator()
}
\value{
a java.object instance
}
\description{
It creates a java.object that can compute the estimate of dissimilarity
indices of Simpson, Sorensen and nestedness.
}
