% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghgVarianceEstimatorsFunctions.R
\name{createEstimate}
\alias{createEstimate}
\title{Construct instance of the estimate class.}
\usage{
createEstimate(mean, variance, dist = "normal")
}
\arguments{
\item{mean}{the point estimate}

\item{variance}{the variance estimate}

\item{dist}{the distribution of the estimator ("normal", "uniform", "lognormal", or "unknown"). By default
the distribution is set to "normal"}
}
\value{
an instance of the estimate class
}
\description{
This function returns an estimate that can later be used in the product of estimates in the context of
greenhouse gas emissions.
}
