% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarityIndicesFunctions.R
\name{createSample}
\alias{createSample}
\title{Create a Sample}
\usage{
createSample(dataSet, plotIdField, speciesIdField)
}
\arguments{
\item{dataSet}{a data.frame object}

\item{plotIdField}{a character string which stands for the field that contains the plot ids.}

\item{speciesIdField}{a character string which stands for the field that contains the species names.}
}
\value{
a java.object instance that represents a Map object in the Java environment
}
\description{
It creates a sample from which the dissimilarity indices can be estimated.
}
