% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{cst_raster}
\alias{cst_raster}
\title{Return or save a 2D raster from a 3D cstdata netcdf file dataset using an
aggregating function.}
\usage{
cst_raster(
  file,
  agg_fun = "mean",
  save = FALSE,
  save_path,
  overwrite = FALSE,
  navalue = -9999,
  compress = FALSE,
  ncores = 1
)
}
\arguments{
\item{file}{A single file path to a netcdf file retrieved by `cstdata()`}

\item{agg_fun}{The aggregating function to applied to the file variable.
(e.g., "mean"). (character)}

\item{save}{Whether to save to file. If `FALSE`, a RasterLayer object will be
returned. Defaults to `FALSE`. (logical)}

\item{save_path}{If saving, the path to the output file. If left blank, the
save_path will be near copy of the input file path, replacing "daily.nc" with
'<agg_fun>.tif'. (character)}

\item{overwrite}{If saving, whether to overwrite an existing file. Defaults
to FALSE. (logical)}

\item{navalue}{If saving, the number to use for NAN values. Defaults to
-9999. (numeric)}

\item{compress}{If saving, whether to compress the resulting file using
using the LZW method. Defaults to FALSE. (logical)}

\item{ncores}{Number of cores to use (int)}
}
\value{
A RasterLayer object resulting the aggregation of all time periods at
each cell according to the aggregating function.
}
\description{
Return or save a 2D raster from a 3D cstdata netcdf file dataset using an
aggregating function.
}
\examples{
\dontrun{
file_ref <- cstdata(park = "Acadia National Park", 
                    parameters = "pr", 
                    years = c(2020, 2021), 
                    models = "CCSM4", 
                    scenarios = "rcp85")
r <- cst_raster(file_ref$path[1],agg_fun = "min")
}

}
