% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghgVarianceEstimatorsFunctions.R
\docType{class}
\name{estimate-class}
\alias{estimate-class}
\alias{estimate}
\title{A class that contains a point and a variance estimates for greenhouse gas emissions.}
\description{
A class that contains a point and a variance estimates for greenhouse gas emissions.
}
