% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{get_reference}
\alias{get_reference}
\title{Return a reference object for the arguments associated with a data set used in cft.}
\usage{
get_reference(dataset)
}
\arguments{
\item{dataset}{A data set from which to download data. Currently available data
sets include 'MACA' and 'GridMET'. (character)}
}
\value{
A reference class object containing available input parameters for `get_gridmet()`
or `get_maca()` and reference information about each.
}
\description{
Return a reference object for the arguments associated with a data set used in cft.
}
\examples{
\dontrun{
file_ref <- get_reference("maca")
print(file_ref$parameters)
}

}
