% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassFunctions.R
\name{getAboveGroundBiomassMg}
\alias{getAboveGroundBiomassMg}
\title{Calculate the aboveground biomass of a list of trees}
\usage{
getAboveGroundBiomassMg(speciesCode, dbhCm, heightM = NULL)
}
\arguments{
\item{speciesCode}{the species code (see the getAbovegroundBiomassSpeciesList function)}

\item{dbhCm}{diameter at breast height (cm)}

\item{heightM}{tree height (m)}
}
\value{
the aboveground biomass (Mg of dry matter)
}
\description{
Calculate the aboveground biomass of a list of trees
}
