% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassFunctions.R
\name{getAbovegroundBiomassSpeciesList}
\alias{getAbovegroundBiomassSpeciesList}
\title{Provide the list of species codes for Lambert et al.'s model.}
\usage{
getAbovegroundBiomassSpeciesList()
}
\description{
Provide all of the possible species code that can be used with the
getAboveGroundBiomassMg function.
}
