% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarityIndicesFunctions.R
\name{getBaselgaDissimilarityIndices}
\alias{getBaselgaDissimilarityIndices}
\title{Calculates Baselga's Dissimilarity Indices}
\usage{
getBaselgaDissimilarityIndices(dissimilarityEstimator, sample)
}
\arguments{
\item{dissimilarityEstimator}{a java.object that can compute or estimate the dissimilarity indices. It
should be generated using the createDissimilarityIndicesEstimator function}

\item{sample}{a java.object instance that stands for a Map in the Java environment. It should be
generated using the createSample function}
}
\value{
a data.frame object with the estimated dissimilarity indices and their standard errors
}
\description{
It computes Baselga's multiple-site dissimilarity indices of Simpson, Sorensen and
nestedness from a set of assemblages.
}
