% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarityIndicesFunctions.R
\name{getDissimilarityEstimates}
\alias{getDissimilarityEstimates}
\title{Estimate Dissimilarity Indices}
\usage{
getDissimilarityEstimates(dissimilarityEstimator, sample, populationSize)
}
\arguments{
\item{dissimilarityEstimator}{a java.object that can estimate the dissimilarity indices. It
should be generated using the createDissimilarityIndicesEstimator function}

\item{sample}{a java.object instance that stands for a Map in the Java environment. It should be
generated using the createSample function}

\item{populationSize}{the number of units (plots) in the population}
}
\value{
a data.frame object with the estimated dissimilarity indices and their standard errors
}
\description{
It estimates the dissimilarity indices of Simpson, Sorensen and
nestedness from a sample.
}
