% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumeFunctions.R
\name{getMerchantableVolumeDm3}
\alias{getMerchantableVolumeDm3}
\title{Calculate the volume of a list of trees}
\usage{
getMerchantableVolumeDm3(speciesCode, dbhCm, heightM, overbark = F)
}
\arguments{
\item{speciesCode}{either the Latin name or a three-character
species code (e.g. "SAB" or "EPB")}

\item{dbhCm}{diameter at breast height (cm)}

\item{heightM}{tree height (m)}

\item{overbark}{true to get the overbark volume (set to FALSE by default)}
}
\value{
the underbark merchantable volume (dm3)
}
\description{
Calculate the volume of a list of trees
}
