% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumeFunctions.R
\name{getMerchantableVolumeSpeciesList}
\alias{getMerchantableVolumeSpeciesList}
\title{Provide the list of species codes}
\usage{
getMerchantableVolumeSpeciesList(latinName = F)
}
\arguments{
\item{latinName}{a logical TRUE to get the eligible Latin names or
FALSE to get the eligible species codes.
Provide all of the possible species code that can be used with the
getUnderbarkMerchantableVolumeDm3 function.}
}
\description{
Provide the list of species codes
}
