% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumeFunctions.R
\name{getUnderbarkMerchantableVolumeDm3}
\alias{getUnderbarkMerchantableVolumeDm3}
\title{Calculate the volume of a list of trees}
\usage{
getUnderbarkMerchantableVolumeDm3(speciesCode, dbhCm, heightM)
}
\arguments{
\item{speciesCode}{a three-slot species code (e.g. "SAB" or "EPB")}

\item{dbhCm}{diameter at breast height (cm)}

\item{heightM}{tree height (m)}
}
\value{
the underbark merchantable volume (dm3)
}
\description{
Calculate the volume of a list of trees
}
