% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\docType{data}
\name{grid_reference}
\alias{grid_reference}
\title{Attributes}
\format{An object of class \code{Grid_Reference} of length 1.}
\usage{
grid_reference
}
\arguments{
\item{grid_reference}{Reference object containing geographical coordinate information of the full grid in which the data is provided.}
}
\description{
Attributes
}
\keyword{datasets}
