% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_gridmet.R
\name{GridMet}
\alias{GridMet}
\title{gridMET Subset generator}
\description{
An R6 class data subset generator for the gridMET dataset.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_url}}{Base URL for data queries.}

\item{\code{catalog_url}}{URL to catalog of all available full.}

\item{\code{arg_ref}}{Argument options and parameter attributes.}

\item{\code{grid_ref}}{Grid attributes.}

\item{\code{project_dir}}{The local directory in which to save files.}

\item{\code{verbose}}{Print verbose output.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{server_open}}{Check access to server.}

\item{\code{existing_aois}}{Check previously downloaded AOI rasters in currently set project directory.}

\item{\code{project_contents}}{Check existing climate files for the currently set aoi.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GridMet$new()}}
\item \href{#method-print}{\code{GridMet$print()}}
\item \href{#method-set_aoi}{\code{GridMet$set_aoi()}}
\item \href{#method-get_subset}{\code{GridMet$get_subset()}}
\item \href{#method-clone}{\code{GridMet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize GridMet Object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridMet$new(project_dir = ".", verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_dir}}{The local directory in which to save files. (character)}

\item{\code{verbose}}{Print verbose output. (logical)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print GridMet attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridMet$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_aoi"></a>}}
\if{latex}{\out{\hypertarget{method-set_aoi}{}}}
\subsection{Method \code{set_aoi()}}{
Set the area of interest for the next query.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridMet$set_aoi(shp_path, area_name, park)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shp_path}}{A path to a shapefile with which to clip the resulting data
sets. This path may be to a local .shp file or a url to a zipped remote
file. If this option is used, leave the `park` argument empty or set
to NA. (character)}

\item{\code{area_name}}{If a shapefile path is used, provide a name to use as a
reference to the location. This will be used in file names, attributes, and
directories. (character)'}

\item{\code{park}}{The name of a national park (e.g., "Yellowstone National
Park", "Yellowstone Park", "Yellowstone", "yellowstone", etc.). The user may
use this option in place of a shapefile path. In this case, leave the
shp_path argument empty or set to NA. (character)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_subset"></a>}}
\if{latex}{\out{\hypertarget{method-get_subset}{}}}
\subsection{Method \code{get_subset()}}{
Retrieve subset of GridMet data with currently set parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridMet$get_subset(models, parameters, scenarios, years, ncores = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{models}}{A list of global circulation models to download. Only available
for GCMs. If left empty all available models will be downloaded. A list of
currently available models is available under cft::get_reference("maca")$models.
(vector)}

\item{\code{parameters}}{A list of climate parameters to download. If left empty
all available parameters will be downloaded. A list of available of models is
available under cft::get_reference(<dataset>)$parameters. (vector)}

\item{\code{scenarios}}{A list of representative concentration pathways (rcps) to
download. If left empty all available rcps will be downloaded. A list of
available of rcps is available under cft::get_reference(<dataset>)$scenarios.
(vector)}

\item{\code{years}}{The first and last years of the desired period. (vector)}

\item{\code{ncores}}{The number of cpus to use, which defaults to 1. (numeric)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridMet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
