% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\docType{data}
\name{gridmet_reference}
\alias{gridmet_reference}
\title{Attributes}
\format{
An object of class \code{GridMET_Reference} of length 1.
}
\usage{
gridmet_reference
}
\arguments{
\item{gridmet_reference}{object containing information about input arguments
for the get_gridmet.}
}
\description{
Attributes
}
\keyword{datasets}
