% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate_setup.R
\name{install_py_deps}
\alias{install_py_deps}
\title{Install python dependencies}
\usage{
install_py_deps(conda = TRUE)
}
\arguments{
\item{conda}{Whether to install dependencies with conda (logical).}
}
\description{
Installs python dependencies of cstdata.
}
