% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_maca.R
\docType{data}
\name{Maca}
\alias{Maca}
\title{MACA Subset generator}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Maca
}
\arguments{
\item{local_dir}{The local directory in which to save files. (character)}

\item{verbose}{Print verbose output. (logical)}

\item{shp_path}{A path to a shapefile with which to clip the resulting data
sets. This path may be to a local .shp file or a url to a zipped remote
file. If this option is used, leave the `park` argument empty or set
to NA. (character)}

\item{area_name}{If a shapefile path is used, provide a name to use as a
reference to the location. This will be used in file names, attributes, and
directories. (character)'}

\item{park}{The name of a national park (e.g., "Yellowstone National
Park", "Yellowstone Park", "Yellowstone", "yellowstone", etc.). The user may
use this option in place of a shapefile path. In this case, leave the
shp_path argument empty or set to NA. (character)}

\item{models}{A list of global circulation models to download. Only available
for GCMs. If left empty all available models will be downloaded. A list of
currently available models is available under cft::get_reference("maca")$models.
(vector)}

\item{parameters}{A list of climate parameters to download. If left empty
all available parameters will be downloaded. A list of available of models is
available under cft::get_reference(<dataset>)$parameters. (vector)}

\item{scenarios}{A list of representative concentration pathways (rcps) to
download. If left empty all available rcps will be downloaded. A list of
available of rcps is available under cft::get_reference(<dataset>)$scenarios.
(vector)}

\item{ncores}{The number of cpus to use, which defaults to 1. (numeric)}

\item{years}{The first and last years of the desired period. (vector)}
}
\description{
An R6 class data subset generator for the MACA dataset.

Initialize Maca Object

Print Maca attributes

Set the area of interest for the next query.

Retrieve subset of MACA data with currently set parameters.
}
\section{Fields}{

\describe{
\item{\code{base_url}}{Base URL for data queries.}

\item{\code{catalog_url}}{URL to catalog of links for all available datasets.}

\item{\code{arg_ref}}{Argument options and parameter attributes.}

\item{\code{grid_ref}}{Grid attributes.}

\item{\code{local_dir}}{The local directory in which to save files.}

\item{\code{verbose}}{Print verbose output.}

\item{\code{server_open}}{Check access to server.}

\item{\code{existing_aois}}{Check previously downloaded AOI rasters in currently set project directory.}

\item{\code{project_contents}}{Check existing climate files for the currently set aoi.}
}}

\keyword{datasets}
