% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\docType{data}
\name{maca_reference}
\alias{maca_reference}
\title{Attributes}
\format{
An object of class \code{MACA_Reference} of length 1.
}
\usage{
maca_reference
}
\arguments{
\item{maca_reference}{object containing information about input marguments
for get_maca.}
}
\description{
Attributes
}
\keyword{datasets}
