% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghgVarianceEstimatorsFunctions.R
\docType{class}
\name{monteCarloEstimate-class}
\alias{monteCarloEstimate-class}
\alias{monteCarloEstimate}
\title{A class that extends the estimate class. In addition to the fields
of the super class, it also contains a list of realizations.}
\description{
A class that extends the estimate class. In addition to the fields
of the super class, it also contains a list of realizations.
}
