% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghgVarianceEstimatorsFunctions.R
\name{multiplyEstimates}
\alias{multiplyEstimates}
\title{Computes the products of several estimates under different methods.}
\usage{
multiplyEstimates(..., factor = 1, method = "Goodman", realMC = 5000)
}
\arguments{
\item{...}{a series of instances of the estimate class}

\item{factor}{a multiplicative factor. The mean and the variance of the resulting
estimate instance are multiplied by factor and factor^2, respectively. By default
this factor is set to 1.}

\item{method}{either "Goodman", "Naive", "Propagation","MonteCarlo", or "RescaledMonteCarlo". The
last two are based on resampling method and the resulting estimate is then a monteCarloEstimate instance
instead of an estimate instance}

\item{realMC}{the number of realizations in cases the method is "MonteCarlo" or "RescaledMonteCarlo"}
}
\value{
an instance of the estimate or the monteCarloEstimate class
}
\description{
Computes the products of several estimates under different methods.
}
