% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghgVarianceEstimatorsFunctions.R
\name{plotEstimate}
\alias{plotEstimate}
\title{Plots the distribution of an estimate.}
\usage{
plotEstimate(estimate)
}
\arguments{
\item{estimate}{an instance of the estimate class}
}
\description{
Plots the distribution of an estimate.
}
