% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatterplot}
\usage{
scatterplot(file_df, var1 = "tasmax", var2 = "pr", agg_fun = "mean",
  difference = TRUE, target_period = c(2080, 2099),
  reference_period = c(1990, 2019), scenarios = c("rcp45", "rcp85"))
}
\arguments{
\item{file_df}{Description coming soon!}

\item{var1}{Description coming soon!}

\item{var2}{Description coming soon!}

\item{agg_fun}{Description coming soon!}

\item{difference}{Description coming soon!}

\item{target_period}{Description coming soon!}

\item{reference_period}{Description coming soon!}

\item{scenarios}{Description coming soon!}
}
\description{
Scatterplot
}
