% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarityIndicesFunctions.R
\docType{data}
\name{subsetUrbanEnvironmentNancy}
\alias{subsetUrbanEnvironmentNancy}
\title{Data from Urban Environments in Nancy, France}
\format{
An object of class \code{data.frame} with 287 rows and 8 columns.
}
\usage{
data(subsetUrbanEnvironmentNancy)
}
\description{
A subset of a two-phase inventory that was carried out
in Nancy, Eastern France, in 2015
}
\examples{
data(subsetUrbanEnvironmentNancy)
}
\keyword{datasets}
