% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghgVarianceEstimatorsFunctions.R
\name{summary.estimate}
\alias{summary.estimate}
\title{Extends the summary function to the estimate class.}
\usage{
\method{summary}{estimate}(object, ...)
}
\arguments{
\item{object}{an instance of the estimate class}

\item{...}{useless parameters}
}
\value{
a character string with the mean, the variance and the distribution of the estimate
}
\description{
Extends the summary function to the estimate class.
}
