% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_data.R
\name{available_data}
\alias{available_data}
\title{Available data}
\usage{
available_data(
  web_link = "https://cida.usgs.gov/thredds/dodsC/macav2metdata_daily_future",
  verbose = TRUE
)
}
\arguments{
\item{web_link}{A web link to the api data source}

\item{verbose}{Should the api outputs be combined with internal metadata}
}
\value{
Data.frame of requested data
}
\description{
Available data
}
