% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cft_df.R
\name{cft_df}
\alias{cft_df}
\title{Generate a climate data frame of spatial averages from the file
reference object generated by cftdata. Averages represent all
grid cells within or touching the area of interest provided to cftdata.}
\usage{
cft_df(file_reference, ncores = 1)
}
\arguments{
\item{file_reference}{A file reference data frame generated by `cftdata()`}

\item{ncores}{Number of cores to use (int)}
}
\value{
A tibble of climate data averaged spatially over an area of interest.
}
\description{
Generate a climate data frame of spatial averages from the file
reference object generated by cftdata. Averages represent all
grid cells within or touching the area of interest provided to cftdata.
}
\examples{
\dontrun{
file_ref <- cftdata(park = "Acadia National Park", 
                    parameters = "pr", 
                    years = c(2020, 2021), 
                    models = "CCSM4", 
                    scenarios = "rcp85")
df <- cft_df(file_ref, ncores = parallel::detectCores())
}

}
