% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cst_df.R
\name{cst_df}
\alias{cst_df}
\title{Generate a climate data frame}
\usage{
cst_df(file_reference, cores = 1)
}
\arguments{
\item{file_reference}{A file reference data frame generated by `cstdata()`}

\item{cores}{Number of cores to use (int)}
}
\value{
A tibble of climate data averaged spatially over an area of interest.
}
\description{
Generate a climate data frame
}
\examples{
\dontrun{
file_ref <- cstdata(park = "Acadia National Park", 
                    parameters = "pr", 
                    years = c(2020, 2021), 
                    models = "CCSM4", 
                    scenarios = "rcp85")
df <- cst_df(file_ref, cores = parallel::detectCores())
}

}
