% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cstdata.R
\name{cstdata}
\alias{cstdata}
\title{Climate Scenario Toolkit Data}
\usage{
cstdata(shp_path = NA, area_name = NA, national_park = NA,
  start_year = 1950, end_year = 2099, store_locally = TRUE,
  local_dir = tempdir(), store_remotely = FALSE,
  aws_config_dir = "~/.aws", models = "all", parameters = "all",
  verbose = TRUE)
}
\arguments{
\item{shp_path}{A path to a shapefile with which to clip the resulting data
sets. This path may be to a local .shp file or a url to a zipped remote
file. If this option is used, leave the national_park argument empty or set
to NA. (character)}

\item{area_name}{If a shapefile path is provided, provide a name to use as a
reference to the location. This will be used in file names and directories. 
(character)'}

\item{national_park}{The name of a national park (e.g., "Yellowstone National
Park". The user may use this option in place of a shapefile path. In this
case, leave the shp_path argument empty or set to NA. (character)}

\item{start_year}{The first year of the desired period. (integer)}

\item{end_year}{The last year of the desired period. (integer)}

\item{store_locally}{If `TRUE` this function will store the results in a
local directory as NetCDF files. This options may be set to `FALSE` to save
disk space, but the `store_remotely` option must be set to `TRUE` in this
case. (logical)}

\item{local_dir}{The local directory in which to save files if
`store_locally` is set to `TRUE`. (character)}

\item{store_remotely}{If `TRUE` this function will store the results in an
Amazon Web Services S3 bucket. This will require the user to store an S3
configuration file on their local machine. (logical)}

\item{aws_config_dir}{The local directory in which to save the configuration
file needed for storing data in an AWS S3 bucket. If the file is not yet
present in this directory, the user will be prompted for the information
needed to build the file. (character)}

\item{verbose}{Print verbose output. (logical)}
}
\description{
Retrieves subsetted data of climate future scenarios for National Parks in
the Contiguous United States. This data is downscaled using the Multivariate
Adaptive Constructed Analogs (MACA) technique.
}
\details{
This package retrieves daily gridded data sets of General Climate Models
(GCM) clipped to specified National Parks. Each of these data sets represent
a single GCM, climate variable and Representative Concentration Pathway (RCP)
from 1950 to 2099. The 1950 to 2005 portion of this time period represents
historical data while the 2006 to 2099 portion represents modeled data. These
can be stored as NetCDF files either locally or on an Amazon Web Service S3
bucket. The original data sets may be found at
\url{ http://thredds.northwestknowledge.net:8080/thredds/reacch_climate_
CMIP5_aggregated_macav2_catalog.html}

Production Notes:

- The use of reticulate may enable us to use Zarr arrays, which are
  accessible directly from an s3 bucket.
- Xarray 0.14 will change the behaviour of open_mfdataset, 0.13 is giving 
  deprecation warnings. Perhaps we pin xarray to 0.12 (same behavior,
  no warnings).
}
