% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cstdata.R
\name{cstdata}
\alias{cstdata}
\title{Get climate future scenarios for the Contiguous United States using the
Multivariate Adaptive Constructed Analogs (MACA) technique.}
\usage{
cstdata(parkname = "Acadia National Park")
}
\arguments{
\item{parkname}{The name of the national park for which to download data
(character), e.g., "Acadia National Park".}
}
\description{
This package retrieves daily gridded data sets of General Climate Models
(GCM) clipped to specified National Parks. Each of these data sets represent
a single GCM, climate variable and Representative Concentration Pathway (RCP)
from 1950 to 2099. The 1950 to 2005 portion of this time period represents
historical data while the 2006 to 2099 portion represents modeled data. These
can be stored as NetCDF files either locally or on an Amazon Web Service S3
bucket. The original data sets may be found at
\url{ http://thredds.northwestknowledge.net:8080/thredds/reacch_climate_
CMIP5_aggregated_macav2_catalog.html}
}
\details{
Production Notes:
\itemize{
\item The use of reticulate may enable us to use Zarr arrays, which are accessible
directly from an s3 bucket.
\item There is also a non-aggregated catalog:
\url{http://thredds.northwestknowledge.net:8080/thredds/reacch_climate_CMIP5_
  macav2_catalog2.html}
These files come in 5 year chunks but have daily windspeed (was) and monthly
potential evapotranspiration (PotEvap) in addition to the other variables.
Imtiaz would like the monthl PotEvap.
\item The Vapor Pressure Deficit products in the aggregated catalog has an issue
with fill values. I've seen this before and it is easily fixed by appending
"#fillmistmatch" to the end of the query, which I will do when I get the
computer back from a test run.
\item This currently saves everything to disk first, which can be a problem. The
alternative is to save each file to a tempfile and overwrite. We'll have to
be careful when parallelizing. Now, Imtiaz has expressed interest in the
ability to save locally. For small parks this is fine, so perhaps an option.
\item For Death Valley using 7 cores with 15.6 GB of RAM and 2GB of SWAP, it
crashed about 275 files in.
}
}
