% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_reference.R
\name{get_reference}
\alias{get_reference}
\title{Get Climate Futures Toolbox Argument Reference object.}
\usage{
get_reference(dataset)
}
\arguments{
\item{dataset}{The name of the data source from which to retrieve options for
cftdata. Defaults to "maca". (character)
\itemize{
  \item "maca" = Multivariate Adaptive Constructed Analogs
  \item "gridmet" = Gridded Surface Meteorological Dataset
  \item "grid" = Shared grid attributes
}}
}
\value{
A reference class object containing argument options of parameter
 attributes for the chosen dataset.
}
\description{
Retrieve a cftdata argument reference object for a given dataset.
}
