% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarityIndicesFunctions.R
\name{getDissimilarityEstimates}
\alias{getDissimilarityEstimates}
\title{Estimate Dissimilarity Indices}
\usage{
getDissimilarityEstimates(dissimilarityEstimator, sample, populationSize)
}
\arguments{
\item{dissimilarityEstimator}{a java.object that can estimate the dissimilarity indices. It
should be generated using the createDissimilarityIndicesEstimator function}

\item{sample}{a java.object instance that stands for a Map in the Java environment. It should be
generated using the createSample function}

\item{populationSize}{the number of units (plots) in the population}
}
\value{
a data.frame object with the estimated dissimilarity indices and their standard errors
}
\description{
This function computes an estimate of the dissimilarity indices of Simpson, Sorensen and
nestedness from a sample.
}
\details{
The processing of this function is done in Java using the J4R package. In order to use this function,
the user must first run the createDissimilarityIndicesEstimator function which creates a Java object that
can estimate the dissimilarity. The result of the createDissimilarityIndicesEstimator function must be
stored in a variable and passed to this function.

A sample of plots with species observation must also be intantiated using the createSample function.
The result of the createSample function is the second parameter of this function.

Finally, the user must provide the population size, that is the number of plots that fit in this population.
}
\examples{

### An example using the subsetUrbanEnvironmentNancy dataset ###

dataReleves <- CFT::subsetUrbanEnvironmentNancy

dissEst <- createDissimilarityIndicesEstimator()
strataList <- unique(dataReleves$Stratum)
output <- NULL
baselga <- NULL
stratum <- strataList[1]

for (stratum in strataList) {
  releve.s <- dataReleves[which(dataReleves$Stratum == stratum),]
  if (stratum == "forest") {
    populationSize <- 3089 * 10000 / (pi * 5^2)
  } else if (stratum == "parking") {
    populationSize <- 501 * 10000 / (pi * 5^2)
  } else {
    populationSize <- 100000
  }

  sample <- createSample(releve.s, "CODE_POINT", "Espece")
  indices <- getDissimilarityEstimates(dissEst, sample, populationSize)
  indices$stratum <- stratum
  output <- rbind(output, indices)
  baselga.s <- getBaselgaDissimilarityIndices(dissEst, sample)
  baselga.s$stratum <- stratum
  baselga <- rbind(baselga, baselga.s)
}
}
\seealso{
createDissimilarityIndicesEstimator

createSample
}
