% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_py_deps.R
\name{install_py_deps}
\alias{install_py_deps}
\title{Install python dependencies}
\usage{
install_py_deps(
  method = "conda",
  conda = "auto",
  envname = "cft",
  restart_session = TRUE,
  ...
)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows. Note also
that since this command runs without privilege the "system" method is
available only on Windows.}

\item{conda}{Path to conda executable (or "auto" to find conda using the
PATH and other conventional install locations).}

\item{envname}{Name of Python environment to install within}

\item{restart_session}{Restart R session after installing (note this will
only occur within RStudio).}

\item{...}{other arguments passed to [reticulate::conda_install()] or
[reticulate::virtualenv_install()].}
}
\description{
Installs python dependencies of cft.
}
\details{
The implementation and interface are adapted from the 
`tensorflow::install_tensorflow()` function.
}
