% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{scatterplot}
\alias{scatterplot}
\title{Climate Scenarios Toolkit Scatterplot}
\usage{
scatterplot(
  file_df = NA,
  agg_fun = "mean",
  var1 = "tasmax",
  var2 = "pr",
  months1 = "all",
  months2 = "all",
  target_period = c(2080, 2099),
  reference_period = c(1990, 2019),
  difference = TRUE,
  scenarios = c("rcp45", "rcp85")
)
}
\arguments{
\item{file_df}{A file reference data frame output from a `cstdata` run.
(data frame)}

\item{agg_fun}{The aggregating function to applied to the model variables
(e.g., "mean"). (character)}

\item{var1}{The x-axis variable, see cstdata::argument_reference
(e.g., "pr"). (character).}

\item{var2}{The y-axis variable, see cstdata::argument_reference
(e.g., "tasmax").  character)}

\item{months1}{The months of the year with which to include in the
aggregation of variable 1 (e.g., c("aug", "sep", "oct")). (vector)}

\item{months2}{The months of the year with which to include in the
aggregation of variable 2 (e.g., c("apr", "may", "jun")). (vector)}

\item{target_period}{The time period within which to apply the aggreating
value function (e.g., c(2050, 2075)). (vector)}

\item{reference_period}{The time period with which to compare the target
period (e.g., c(1980, 2010)). (vector)}

\item{difference}{Whether or not to return values as the difference between
the target and reference periods. Requires an input for `reference_period`.
(logical).}

\item{scenarios}{A vector of representative concetration pathway keys
(e.g., c("rcp45", rcp85")). (vector)}
}
\description{
Renders a scatterplot between two model variables from an assortment of
General Circulation Models. This function uses an user-specified function to
aggregate model outputs within a specified area and time period. Users may
also choose to return differences between the results of that aggregating
function between two time periods. Uses the output of `cstdata` as a
reference to local or remote files and meta data.
}
